refactor: Drawing 모듈 폴더 정리
