+++ server-toss/src/test-setup/setup-mikro-orm-mocks.ts
@@ -16,4 +16,5 @@ jest.mock("@mikro-orm/decorators/legacy", () => ({
   ManyToOne: () => () => undefined,
   ManyToMany: () => () => undefined,
   OneToMany: () => () => undefined,
+  Enum: () => () => undefined,
 }));
